refactor: 단일책임원칙에 인한 클래스 분리
